clean file structure
